refactor: extract core logic and interfaces from index.ts
- Move createConversation and createAudioSession to core.ts
- Extract IWebSocket interface to interfaces/websocket.ts
- Simplify index.ts to only export createConversation
