+++ frontend/src/interfaces/websocket.ts
@@ -1,0 +1,6 @@
+interface IWebSocket {
+    ready(): boolean;
+    send(data: string | ArrayBuffer): void;
+    close(): void;
+    addEventListener(type: string, listener: (evt: any) => void): void;
+}
